fix: refactor routes to use renamed container file
